refactor:change price properties to double

+++ SavingPlatformApplication/Data/Models/Credit.cs
@@ -4,9 +4,9 @@ namespace SavingPlatformApplication.Data.Models
 {
     public class Credit : BaseModel
     {
-        public decimal FinancialResourceAmount { get; set; }
+        public double FinancialResourceAmount { get; set; }
 
-        public decimal PercentInterest { get; set; }
+        public double PercentInterest { get; set; }
 
         public DateTime PaymentPeriod { get; set; }
